prevent rendering if readme is changed
