+++ packages/neoFrontend/src/pages/VizPage/RunContext/changesMD.js
@@ -1,0 +1,5 @@
+import { createChangesChecker } from './createChangesChecker';
+
+// Returns true if the given op changes a .md file
+// that is not bundle.js.
+export const changesMD = createChangesChecker('.md');
