+++ packages/neoFrontend/src/pages/VizPage/RunContext/index.js
@@ -5,6 +5,6 @@ export const RunContext = createContext();
 
 // This context is responsible for figuring out when to re-run the code,
 // and when to recompute and store bundle.js.
-export const RunProvider = ({ fallback, children }) => (
+export const RunProvider = ({ children }) => (
   <RunContext.Provider value={useRun()}>{children}</RunContext.Provider>
 );
